requirements.txt
anyio==4.0.0
async-timeout==4.0.3
asyncpg==0.29.0
fastapi==0.92.0
graphql-core==3.2.3
pydantic==1.10.13
starlette==0.25.0
strawberry-graphql==0.158.0
uvicorn==0.20.0
